Fix wrong vararg type in LFLock_init
This only results in random junk in a tempfile name, so probably
harmless.

+++ core/Lucy/Store/Lock.c
@@ -124,7 +124,8 @@ LFLock_init(LockFileLock *self, Folder *folder, String *name,
     int pid = PID_getpid();
     Lock_init((Lock*)self, folder, name, host, timeout, interval);
     LockFileLockIVARS *const ivars = LFLock_IVARS(self);
-    ivars->link_path = Str_newf("%o.%o.%i64", ivars->lock_path, host, pid);
+    ivars->link_path = Str_newf("%o.%o.%i64", ivars->lock_path, host,
+                                (int64_t)pid);
     return self;
 }
